refactor: US-0 / apply PR feedback
