+++ src/InnoClinic.AccountService/InnoClinic.AccountService.WebAPI/InnoClinic.AccountService.WebAPI.csproj
@@ -1,9 +1,10 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
     <TargetFramework>net8.0</TargetFramework>
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
+    <UserSecretsId>929a6053-36bd-4602-9161-563e42b63f67</UserSecretsId>
   </PropertyGroup>
 
   <ItemGroup>
