+++ src/InnoClinic.AccountService/InnoClinic.AccountService.WebAPI/appsettings.json
@@ -7,6 +7,6 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "AccountDb": "Host=localhost;Port=5432;Database=innoclinic;Username=admin;Password=password"
+    "AccountDb": ""
   }
 }
